+++ src/main/java/com/github/johnmedlockdev/main/modes/interfaces/ModeInterface.java
@@ -1,6 +1,6 @@
 package com.github.johnmedlockdev.main.modes.interfaces;
 
-public interface Mode {
+public interface ModeInterface {
 
     void gatherInput();
 }