+++ src/main/java/com/github/johnmedlockdev/main/routes/Init.java
@@ -1,33 +1,40 @@
 package com.github.johnmedlockdev.main.routes;
 
+import com.github.johnmedlockdev.main.modes.AdviseMode;
 import com.github.johnmedlockdev.main.modes.BatchMode;
 import com.github.johnmedlockdev.main.modes.GenerateMode;
 import com.github.johnmedlockdev.main.modes.ManualMode;
-import com.github.johnmedlockdev.main.modes.PredictMode;
-import com.github.johnmedlockdev.main.modes.interfaces.Mode;
+import com.github.johnmedlockdev.main.modes.interfaces.ModeInterface;
 
 public class Init {
 
     public Init(String[] args) {
         String selection = args[0];
-        logic(selection);
+        System.out.println("Select Mode:");
+        System.out.println("Manual : M || Batch : B || Generate : G || Advise : A");
+        modeFactory(selection);
     }
 
-    private void logic(String selection) {
+    private void modeFactory(String selection) {
         selection = capitalizeInput(selection);
-        Mode mode;
+        ModeInterface mode;
+
         switch (selection) {
             case "Manual":
+            case "M":
                 mode = new ManualMode();
                 break;
             case "Batch":
+            case "B":
                 mode = new BatchMode();
                 break;
             case "Generate":
+            case "G":
                 mode = new GenerateMode();
                 break;
-            case "Predict":
-                mode = new PredictMode();
+            case "Advise":
+            case "A":
+                mode = new AdviseMode();
                 break;
             default:
                 throw new IllegalStateException("Unexpected value: " + selection);