+++ src/main/java/com/github/johnmedlockdev/main/strategy/interfaces/StrategyInterface.java
@@ -1,0 +1,10 @@
+package com.github.johnmedlockdev.main.strategy.interfaces;
+
+import java.util.ArrayList;
+
+public interface StrategyInterface {
+
+    ArrayList<Double> processData();
+
+    void advise();
+}