+++ src/main/java/com/github/johnmedlockdev/main/strategy/interfaces/StrategyTemplate.java
@@ -1,8 +1,0 @@
-package com.github.johnmedlockdev.main.strategy.interfaces;
-
-public interface StrategyTemplate {
-
-    double[] processData();
-
-    void advise();
-}